Handle blank entries in homophones.csv (#1529)
added a line to function update_homophones in homophones.py to remove
words that are empty strings from the list of homophones created from
each line of the CSV file. The issue is that CSV file format is not
standard, and programs like MS Excel try to force each line to have the
same number of "cells" (i.e. commas). This results in almost every line
having on extra comma which is read as a "word" that is an empty string,
and then these empty string "homophones" are matched with other words'
empty string homophones resulting in every word being a homophone of
every other word in the file.

---------

Co-authored-by: Barry Kronenfeld <barrykronenfeld@gmail.com>
Co-authored-by: Nicholas Riley <com-github@sabi.net>
Co-authored-by: Phil Cohen <phillip@phillip.io>