Added setting for selected text timeout (#1517)
The default time out for `clip.capture()` is 0.5s. You run into this
every time you call `selected_text()` on an empty selection. I've been
running with 0.1 for over a year with no problems. That might be a bit
to aggressive for community, but 0.25 shouldn't be any problems. Just in
case I added a setting so people with slower (or faster) machines can
change this.

Fixes #1210